New, more general approach to parse errors and warnings
This introduces InputErrors, which represent positioned errors arising
from problems in the input.  As part of this, the new first-class Pos
objects track positions in input files and serve as the entry-point
for logging warning and logging and raising InputErrors.  A new
context manager allows callers to recover from and bundle InputErrors,
which lets us use the Python exception system to communicate errors in
the input, without limiting us to communicating a single error at a
time.